chore(web-component): remove unused font.style file and extract font injection in utils
